moving intoEvent to core/utils
