+++ src/server/tracer/tests/awsXrayInstrument.spec.ts
@@ -2,7 +2,7 @@ import { Subsegment } from "aws-xray-sdk";
 import instrumentExecutionOnXray from "../awsXrayInstrument";
 import { Event } from "core/events";
 import { buildResponseEventFor } from "../../responseEventBuilder";
-import { intoEvent } from "client/client";
+import { intoEvent } from "core/utils";
 
 const name = "event:test";
 const version = 1;
